refactor: move TemporalFilter to reduce file layout complexity
