--- junifer/preprocess/filter/__init__.py
@@ -1,9 +1,0 @@
-"""Provide imports for filter sub-package."""
-
-# Authors: Federico Raimondo <f.raimondo@fz-juelich.de>
-# License: AGPL
-
-import lazy_loader as lazy
-
-
-__getattr__, __dir__, __all__ = lazy.attach_stub(__name__, __file__)
