--- junifer/preprocess/filter/__init__.pyi
@@ -1,3 +1,0 @@
-__all__ = ["TemporalFilter"]
-
-from .temporalfilter import TemporalFilter
